Address bug causing "TypeError: too many data entries"
This does not fix the bug. The bug is caused by a combination
of the functions combineboundims() and pxfrombounds().
The first does not properly combine overlapping boundaries,
possibly resulting in BoundIm's with repeated boundaries;
these are not properly handled by pxfrombounds(), causing
distorted images and exceeding image dimensions.
The bug is for now solved by preventing the combination
of overlapping images, but the two functions should be
improved anyway, which is noted in TODO comments.